add: graphiql and test. fix: admin ticket delete, graphql template context remove, fix tixckets in query type

+++ app/graphql/mutations/delete_ticket.rb
@@ -8,7 +8,7 @@ module Mutations
       user = context[:current_user]
       ticket = Ticket.find(id)
 
-      raise "Unauthorized" unless ticket.user == user
+      raise "Unauthorized" unless user.admin? || ticket.user == user 
 
       ticket.destroy
       { success: true }
